+++ tests/testthat/test-08-module-power.R
@@ -52,7 +52,7 @@ test_that("power, with LLM", {
 
   module <- "power"
   llm_use(TRUE)
-  llm_model("llama-3.3-70b-versatile")
+  llm_model("groq/llama-3.3-70b-versatile")
 
   # only false positives
   paper <- paper()
@@ -74,7 +74,7 @@ test_that("power, with LLM", {
   mod_output <- module_run(paper, module)
   expect_equal(mod_output$traffic_light, "red")
   expect_equal(nrow(mod_output$table), 1)
-  expect_equal(mod_output$table$sample_size, 15)
+  expect_equal(mod_output$table$sample_size, 30)
   expect_equal(mod_output$table$power, 0.8)
   expect_equal(mod_output$table$effect_size, NA)
   expect_equal(mod_output$table$alpha_level, NA)
@@ -90,7 +90,7 @@ test_that("power, with LLM", {
   expect_equal(mod_output$traffic_light, "red")
   expect_equal(nrow(mod_output$table), 2)
   expect_equal(mod_output$table$statistical_test,
-               c("unpaired t-test", "one-way ANOVA"))
+               c("unpaired t-test", "1-way ANOVA"))
   expect_equal(mod_output$table$sample_size, c(300, 350))
   expect_equal(mod_output$table$alpha_level, c(0.05, NA))
   expect_equal(mod_output$table$power, c(0.8, 0.8))